Replacing calls to bytes.Compare by bytesutil.CompareBytes
