+++ sql/query/expr.go
@@ -102,6 +102,17 @@ func (l LiteralExprList) Eval(stack EvalStack) (document.Value, error) {
 // NamedParam is an expression which represents the name of a parameter.
 type NamedParam string
 
+// Eval looks up for the parameters in the stack for the one that has the same name as p
+// and returns the value.
+func (p NamedParam) Eval(stack EvalStack) (document.Value, error) {
+	v, err := p.extract(stack.Params)
+	if err != nil {
+		return nilLitteral, err
+	}
+
+	return document.NewValue(v)
+}
+
 func (p NamedParam) extract(params []Param) (interface{}, error) {
 	for _, nv := range params {
 		if nv.Name == string(p) {
@@ -114,6 +125,17 @@ func (p NamedParam) extract(params []Param) (interface{}, error) {
 
 // PositionalParam is an expression which represents the position of a parameter.
 type PositionalParam int
+
+// Eval looks up for the parameters in the stack for the one that is has the same position as p
+// and returns the value.
+func (p PositionalParam) Eval(stack EvalStack) (document.Value, error) {
+	v, err := p.extract(stack.Params)
+	if err != nil {
+		return nilLitteral, err
+	}
+
+	return document.NewValue(v)
+}
 
 func (p PositionalParam) extract(params []Param) (interface{}, error) {
 	idx := int(p - 1)
